Add Option for default attribute ommission.
Whil this is not ideal and HTML shouldn't in a way that is affected
by affected by presence of default attributes or lack thereof, but
alas the reality is different. CSS rules written speciically based
on attributes doesn't treat "defaults" differently, for example, CSS
select `input[type="text"]` doesn't to `<input>`.
Real life casualties of this problem is at least MaterializeCSS and
Foundation, but most probably more.