Changed makefile to use Bash
The shebang in a source file does not apply when it is included within another script. This means bash_completion will be interpreted by whichever shell is being used for the makefile, which by default is /bin/sh. On Debian, for example, /bin/sh is dash and so I receive the following error: /bin/sh: 19: cmd/minify/bash_completion: Syntax error: "(" unexpected (expecting "fi").